+++ R/huxreg.R
@@ -12,7 +12,8 @@
 #' @param stars Levels for p value stars. Names of \code{stars} are symbols to use. Set to \code{NULL} to not show stars.
 #' @param bold_signif Where p values are below this number, cells will be displayed in bold. Use \code{NULL} to turn off
 #'   this behaviour.
-#' @param borders Thickness of horizontal borders in appropriate places. Set to 0 for no borders.
+#' @param borders Thickness of inner horizontal borders. Set to 0 for no borders.
+#' @param borders Thickness of outer (top and bottom) horizontal borders. Set to 0 for no borders.
 #' @param note Footnote for bottom cell, which spans all columns. \code{{stars}} will be replaced by a note about
 #'   significance stars. Set to \code{NULL} for no footnote.
 #' @param statistics Summary statistics to display. Set to \code{NULL} to show all available statistics.
@@ -61,6 +62,7 @@ huxreg <- function (
         stars           = c('***' = 0.001, '**' = 0.01, '*' = 0.05),
         bold_signif     = NULL,
         borders         = 0.4,
+        outer_borders   = 0.8,
         note            = '{stars}.',
         statistics      = c('N' = 'nobs', 'R2' = 'r.squared', 'logLik', 'AIC'),
         coefs           = NULL,
@@ -199,7 +201,9 @@ huxreg <- function (
   if (error_pos == 'right') mod_col_headings <- interleave(mod_col_headings, '')
   mod_col_headings <- c('', mod_col_headings)
   result <- rbind(mod_col_headings, cols, sumstats, copy_cell_props = FALSE)
-  result <- set_bottom_border(result, c(1, 1 + nrow(cols), nrow(result)), everywhere, borders)
+  result <- set_bottom_border(result, final(), everywhere, outer_borders)
+  result <- set_top_border(result, 1, everywhere, outer_borders)
+  result <- set_bottom_border(result, c(1, nrow(cols) + 1), -1, borders)
   colnames(result) <- c('names', names_or(models, paste0("model", seq_along(models))))
   if (error_pos == 'right') result <- set_colspan(result, 1, evens, 2)
   align(result)[1, ]    <- 'center'
@@ -210,11 +214,9 @@ huxreg <- function (
     stars <- if (is.null(stars)) '' else paste0(names(stars), ' p < ', stars, collapse = '; ')
     note <- gsub('%stars%', stars, note)
     note <- glue::glue(note)
-    result <- add_footnote(result, note)
+    result <- add_footnote(result, note, border = 0) # borders handled above
     result <- set_wrap(result, final(), 1, TRUE)
     result <- set_align(result, final(), 1, 'left')
-    result <- set_bottom_border(result, final(), everywhere, 0)
-    result <- set_top_border(result, final(), everywhere, borders)
   }
 
   result