+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# huxtable 1.1.0.9000
+
+* New `outer_borders` argument for `huxreg`. This changes default behaviour slightly.
+* New `border` argument for `add_footnote` to choose width of footnote's top border.
+
 # huxtable 1.1.0
 
 * No more ugly autocreated column names.